+++ src/EventSink.h
@@ -3,5 +3,5 @@
 class EventSink {
 public:
   virtual ~EventSink() {}
-  virtual void reportEvent(const char* gameName, const char* key, long value) = 0;
+  virtual void reportEvent(const char* gameName, const char* key, int value) = 0;
 };
