+++ src/NullEventSink.h
@@ -5,5 +5,5 @@
 // No-op today since every scorepad is standalone.
 class NullEventSink : public EventSink {
 public:
-  void reportEvent(const char*, const char*, long) override {}
+  void reportEvent(const char*, const char*, int) override {}
 };
