+++ test/fakes/FakeEventSink.h
@@ -7,9 +7,9 @@ public:
   int callCount = 0;
   std::string lastGameName;
   std::string lastKey;
-  long lastValue = 0;
+  int lastValue = 0;
 
-  void reportEvent(const char* gameName, const char* key, long value) override {
+  void reportEvent(const char* gameName, const char* key, int value) override {
     callCount++;
     lastGameName = gameName;
     lastKey = key;
